+++ nixos/modules/services/databases/redis.nix
@@ -46,7 +46,6 @@ in
       package = mkOption {
         type = types.package;
         default = pkgs.redis;
-        defaultText = "pkgs.redis";
         description = "Which Redis derivation to use.";
       };
 
@@ -66,22 +65,6 @@ in
         type = types.int;
         default = 6379;
         description = "The port for Redis to listen to.";
-      };
-
-      vmOverCommit = mkOption {
-        type = types.bool;
-        default = false;
-        description = ''
-          Set vm.overcommit_memory to 1 (Suggested for Background Saving: http://redis.io/topics/faq)
-        '';
-      };
-
-      openFirewall = mkOption {
-        type = types.bool;
-        default = false;
-        description = ''
-          Whether to open ports in the firewall for the server.
-        '';
       };
 
       bind = mkOption {
@@ -209,14 +192,6 @@ in
 
   config = mkIf config.services.redis.enable {
 
-    boot.kernel.sysctl = mkIf cfg.vmOverCommit {
-      "vm.overcommit_memory" = "1";
-    };
-
-    networking.firewall = mkIf cfg.openFirewall {
-      allowedTCPPorts = [ cfg.port ];
-    };
-
     users.extraUsers.redis =
       { name = cfg.user;
         uid = config.ids.uids.redis;
@@ -226,7 +201,7 @@ in
     environment.systemPackages = [ cfg.package ];
 
     systemd.services.redis_init =
-      { description = "Redis Server Initialisation";
+      { description = "Redis server initialisation";
 
         wantedBy = [ "redis.service" ];
         before = [ "redis.service" ];
@@ -234,13 +209,14 @@ in
         serviceConfig.Type = "oneshot";
 
         script = ''
-          install -d -m0700 -o ${cfg.user} ${cfg.dbpath}
-          chown -R ${cfg.user} ${cfg.dbpath}
+          if ! test -e ${cfg.dbpath}; then
+              install -d -m0700 -o ${cfg.user} ${cfg.dbpath}
+          fi
         '';
       };
 
     systemd.services.redis =
-      { description = "Redis Server";
+      { description = "Redis server";
 
         wantedBy = [ "multi-user.target" ];
         after = [ "network.target" ];
